requirements.txt
numpy
PyYAML
SimpleITK
tqdm
opencv-python
